refactor(beaconevent): extract GenAI builders into genai.go
Continue the converter split. Move the gen_ai.usage token-usage setters
(setUsage*, ApplyTokenUsage) and the GenAI struct builders (GenAIFromAttrs,
GenAIRequestFromAttrs, GenAIResponseFromAttrs, GenAIToolFromAttrs,
GenAIUsageFromAttrs) into a new same-package genai.go.

PopulateCommon and MCPFromAttrs stay in converter.go (core conversion).
Move-only, behavior-preserving: converter.go drops to ~984 lines; tests stay green.
